Flag indicating if typeahead input should be cleared on lost focus
In some cases it is not desirable to clear typeahead input on focus lost if one of the suggested values was not selected. For instance, user may have an inline option to create some entity if no matching value was found. In this case, typeahead input should remain filled on focus lost. On the other hand, sometimes, we want to enforce user to select one of the suggested options that come from already existing entities. In this case typeahead input should be cleared if no valid selection was made.