Let user type the entire value manually
Previously, when user entered entire value manually and left input field in other way than by pressing 'TAB' (e.g. using mouse), elemica-suggest was not behaving like this value was selected. Instead, it was ignoring it and assuming that no valid option was chosen. It's not a very user-friendly approach, especially in case of suggestions that are relatively short. For instance, if we have a typeahead input for selecting country code, there is a higher probability that user will type entire "US" value instead of typing "U" and then using arrows to go to "US" and selecting it. In case when user entered entire value manually, we must ensure that no matter how user leave this field, elemica-suggest will behave the same way. This fix makes sure that if user entered entire value manually and this value is matching one of suggestions, no matter how user leaves the input field (tab/mouse), this value is considered as selected.