Don't invoke afterSelect if user entered nothing
Added conditional check that makes afterSelect not to be invoked when user entered nothing in the text search field. This prevents from the following scenario:
1. User opens page, clicks inside search field
2. User enters nothing, clicks out of search field
3. 'afterSelect' gets invoked <-- !!!

With change introduced in this commit, afterSelect will not get invoked in a scenario described above. However, if user entered anything and then clicked out of text field, afterSelect will be invoked with null param. The same if user cleared previous selection - afterSelect will be invoked with null param.