Don't trigger searches on non-printable characters.
Before, if you were holding a modifier key and let it go after a keydown or up
was processed, an extra search could be triggered. This was particularly
obvious in cases where C-N or C-P were used, since letting go of the control
key would trigger anoter search and thus reset the selection to the top.